Limit saved sectors to a max number of 5
