+++ src/constants/defaults.js
@@ -1,2 +1,3 @@
 export const ROWS = 10;
 export const COLUMNS = 8;
+export const SECTOR_LIMIT = 5;
